requirements.txt
django
channels
psycopg2
djangorestframework
